requirements.txt
python-telegram-bot==13.9
